Fix cyclic dependency.
